Add support for react-helmet-async, fix issue in server rendering
Closes #5

Server rendering failed if console was replayed, because the json-
stringified object was passed to ruby and the console-replay could not
be parsed. Now the js server renderer passes a valid html string and
separates <head> on an html marker comment.

Instead of passing a string from JS to ruby, it could be possible to
return a real object, but this needs to be supported by react-rails'
ServerRenderer (maybe a more customizable approach?).